pyproject.toml
[tool.poetry]
name = "api-online-shop"
version = "0.1.0"
description = ""
authors = ["SRubtsov <frexrub@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
